// wcs-ss-lib/src/Enums.ts
export enum Role {
    Leader = "Leader",
    Follower = "Follower",
    Switch = "Switch"
}

export enum CompetitionType {
    JnJ = "Jack & Jill",
    Strictly = "Strictly"
}

export enum Division {
    Newcomer = "Newcomer",
    Novice = "Novice",
    Intermediate = "Intermediate",
    Advanced = "Advanced",
    AllStar = "All Star",
    Champion = "Champion",
    Open = "Open"
}

export enum Round {
    Prelims = "Prelims",
    Quarterfinals = "Quarterfinals",
    Semifinals = "Semifinals",
    Finals = "Finals"
}

export enum CallbackScore {
    Yes = "Yes",
    No = "No",
    Alternate1 = "Alt1",
    Alternate2 = "Alt2",
    Alternate3 = "Alt3",
    Unscored = ""
}

export enum Tier {
    NoTier,
    Tier1,
    Tier2,
    Tier3,
    Tier4,
    Tier5,
    Tier6,
}